destroy user session on logout
